requirements.txt
Django>=5.0.4
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
dj-database-url>=2.1.0
psycopg[binary,pool]>=3.1.18
Pillow>=10.2.0
